# Newswires backend: improve search performance