<!-- See https://github.com/guardian/recommendations/blob/main/pull-requests.md for recommendations on raising and reviewing pull requests. -->

## What does this change?

Count total records in separate query:
- The count query running separately; it does not force PostgreSQL to scan all matching rows for each result.
- The data query executes faster because it no longer calculates COUNT(*) OVER() for every row.

## How to test

Run: `EXPLAIN ANALYZE <query>`

### Low performance query:
```
 explain analyze SELECT COUNT(*) OVER() AS total_count, 
   fm.id as i_on_fm,
   fm.external_id as ei_on_fm,
   fm.ingested_at as ia_on_fm,
   fm.supplier as s_on_fm,
   fm.content as c_on_fm
 , ts_headline('english', fm.content->>'body_text', websearch_to_tsquery('english', 'london')) AS h_on_fm
 FROM fingerpost_wire_entry fm
 WHERE websearch_to_tsquery('english', 'london') @@ fm.combined_textsearch
 ORDER BY fm.ingested_at DESC
 LIMIT 30;
                                                                                    QUERY PLAN                                                                                    
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.42..939.65 rows=30 width=500) (actual time=57803.566..57825.948 rows=30 loops=1)
   ->  WindowAgg  (cost=0.42..231708.42 rows=7401 width=500) (actual time=57803.564..57825.939 rows=30 loops=1)
         ->  Index Scan Backward using ingested_at_index on fingerpost_wire_entry fm  (cost=0.42..229747.16 rows=7401 width=460) (actual time=0.130..57777.413 rows=6112 loops=1)
               Filter: ('''london'''::tsquery @@ combined_textsearch)
               Rows Removed by Filter: 123664
 Planning Time: 0.194 ms
 Execution Time: 57826.418 ms
(7 rows)
```

 **Total Execution Time**: 57826.418 ms

The window function (COUNT(*) OVER() AS total_count) requires PostgreSQL to evaluate all matching rows before applying the limit. This prevents optimisations like stopping early once 30 rows are found.

### Improvement performance queries:

Total count:
```
explain analyze SELECT COUNT(*)
 FROM fingerpost_wire_entry fm
 WHERE websearch_to_tsquery('english', 'london') @@ fm.combined_textsearch;
                                                                  QUERY PLAN                                                                   
-----------------------------------------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=19429.97..19429.98 rows=1 width=8) (actual time=14.130..14.131 rows=1 loops=1)
   ->  Bitmap Heap Scan on fingerpost_wire_entry fm  (cost=1722.84..19414.97 rows=5999 width=0) (actual time=4.384..13.758 rows=5990 loops=1)
         Recheck Cond: ('''london'''::tsquery @@ combined_textsearch)
         Heap Blocks: exact=5615
         ->  Bitmap Index Scan on combined_textsearch_idx  (cost=0.00..1721.34 rows=5999 width=0) (actual time=3.632..3.633 rows=6381 loops=1)
               Index Cond: (combined_textsearch @@ '''london'''::tsquery)
 Planning Time: 2.844 ms
 Execution Time: 14.251 ms
(8 rows)
```

Data:
```
explain analyze SELECT 
   fm.id as i_on_fm,
   fm.external_id as ei_on_fm,
   fm.ingested_at as ia_on_fm,
   fm.supplier as s_on_fm,
   fm.content as c_on_fm
 , ts_headline('english', fm.content->>'body_text', websearch_to_tsquery('english', 'london')) AS h_on_fm
 FROM fingerpost_wire_entry fm
 WHERE websearch_to_tsquery('english', 'london') @@ fm.combined_textsearch
 ORDER BY fm.ingested_at DESC
 LIMIT 30;
                                                                              QUERY PLAN                                                                               
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.42..1152.98 rows=30 width=500) (actual time=1.066..18.748 rows=30 loops=1)
   ->  Index Scan Backward using ingested_at_index on fingerpost_wire_entry fm  (cost=0.42..230475.20 rows=5999 width=500) (actual time=1.065..18.742 rows=30 loops=1)
         Filter: ('''london'''::tsquery @@ combined_textsearch)
         Rows Removed by Filter: 326
 Planning Time: 0.311 ms
 Execution Time: 18.785 ms
(6 rows)
```

 **Total Execution Time**: 14.251 ms + 8.785 ms

## How can we measure success?

A simple text search used to take 57s to complete; with the improvement, it takes 897ms.

## Have we considered potential risks?

<!-- What are the potential risks and how can they be mitigated? Does an error require an alarm? Should user help, infosec, or legal be informed of this change? Is private information guarded? Do we need to add anything in the backlog? -->

## Images

<!-- Usually only applicable to UI changes, what did it look like before and what will it look like after? -->

## Accessibility

<!-- Usually only applicable to UI changes, check the boxes if you are satisfied that your changes pass these tests -->

-   [ ] [Tested with screen reader](https://github.com/guardian/accessibility/blob/main/people-and-technology/03-visual.md#screen-reader)
-   [ ] [Navigable with keyboard](https://github.com/guardian/accessibility/blob/main/people-and-technology/02-physical.md#keyboard)
-   [ ] [Colour contrast passed](https://github.com/guardian/accessibility/blob/main/people-and-technology/03-visual.md#contrast)
-   [ ] [The change doesn't use only colour to convey meaning](https://github.com/guardian/accessibility/blob/main/people-and-technology/03-visual.md#use-of-colour)